// streamlit_molstar/pocket/frontend/src/custom-types.ts
export interface PocketData {
    name: string;
    rank: string;
    score: string;
    probability: string;
    center: string[];
    residues: string[];
    surface: string[];
    color?: string;             //color of the pocket, if already generated by rcsb plugin (e.g. "00ff00")
    isVisible?: boolean;        //if the pocket is visible
    avgConservation?: number;   //computed average conservation of the pocket
    avgAlphaFold?: number;      //computed average AlphaFold score of the pocket
}